Split CLI into multiple modules for command groups
