+++ tests/cli_test.py
@@ -310,63 +310,6 @@ class ZAPCliTestCase(unittest.TestCase):
                                               '--engine', 'Oracle Nashorn', '--file-path', 'Foo.js'])
         self.assertEqual(result.exit_code, 1)
 
-    @data(
-        (
-            [
-                {'id': 1, 'value': 'one'},
-                {'id': 5, 'value': 'five'},
-                {'id': 10, 'value': 'ten'}
-            ],
-            [1],
-            [{'id': 1, 'value': 'one'}]
-        ),
-        (
-            [
-                {'id': 1, 'value': 'one'},
-                {'id': 5, 'value': 'five'},
-                {'id': 10, 'value': 'ten'}
-            ],
-            [1, 10],
-            [
-                {'id': 1, 'value': 'one'},
-                {'id': 10, 'value': 'ten'}
-            ]
-        ),
-        (
-            [
-                {'id': 1, 'value': 'one'},
-                {'id': 5, 'value': 'five'},
-                {'id': 10, 'value': 'ten'}
-            ],
-            [4],
-            []
-        ),
-        (
-            [
-                {'id': 1, 'value': 'one'},
-                {'id': 5, 'value': 'five'},
-                {'id': 10, 'value': 'ten'}
-            ],
-            [],
-            [
-                {'id': 1, 'value': 'one'},
-                {'id': 5, 'value': 'five'},
-                {'id': 10, 'value': 'ten'}
-            ]
-        ),
-        (
-            [],
-            [],
-            []
-        )
-    )
-    @unpack
-    def test_filter_by_ids(self, original_list, ids_to_filter, expected_result):
-        """Test the function for filtering a list of dicts by IDs."""
-        result = cli.filter_by_ids(original_list, ids_to_filter)
-
-        self.assertEqual(result, expected_result)
-
     @patch('zapcli.zap_helper.ZAPHelper.xml_report')
     def test_xml_report(self, report_mock):
         """Testing XML report."""
